refactor file structure and add code reference links
