+++ aries-site/src/examples/templates/content-layouts/ContainerDrivenLayout/constants.js
@@ -1,0 +1,32 @@
+import { Text } from 'grommet';
+
+export const CATEGORIES = [
+  {
+    value: 'Machine Learning',
+    label: <Text>Machine Learning</Text>,
+  },
+  {
+    value: 'Big Data',
+    label: <Text>Big Data</Text>,
+  },
+  {
+    value: 'Data Analytics',
+    label: <Text>Data Analytics</Text>,
+  },
+  {
+    value: 'Data Protection',
+    label: <Text>Data Protection</Text>,
+  },
+  {
+    value: 'Database',
+    label: <Text>Database</Text>,
+  },
+  {
+    value: 'Developer Tools',
+    label: <Text>Developer Tools</Text>,
+  },
+  {
+    value: 'Monitoring',
+    label: <Text>Monitoring</Text>,
+  },
+];
